Updates to saveUser to accomodate MariaDB datetime values
The original value of `NOW()` was being interpreted as a literal string value
as opposed to being treated as a function call, this was causing user creation
and updates to fail due to an incompatible value being provided for the
`password_last_updated` column. Replacing this w/ a php call to `date` means
that we provide a valid value in the form `Y-m-d H:i:s`.

**NOTE: there will be a follow on PR that refactors the `getUpdateQuery` and
`getInsertQuery` functions into a single `getQuery` function that greatly
simplifies & makes more readable the code responsible for generating the
required SQL Statement for creating / updating a user.**