Updating ContentType to follow finfo
The content_type reported by PHP7 for xml is `text/xml` while in PHP8 it's
reported as `text/xml;charset=UTF-8`. There are two ways we could handle this,
one is with the changes included in this commit. The other is that we can update
`ExportBuilder.php::$supported_formats["xml"]["render_as"]` to include
`charset=UTF-8`.