Update to use ETLOverseer instead of Actions
This change was necessitated by the DBModelTest tests / ETL actions failing to
create the `test` db as expected. After further research I found that this was
due to the action being directory constructed / executed vs. utilizing the
EtlOverseer. With this change, actions that specify that a schema should be
created if one does not exist will operate as expected.