Fix data encoding in etlv2 pdoIngestor
The data encoding in the mysql LOAD DATA INFILE mechanism
in the pdoIngestor was incorrect. It did not properly encode
the data in the cases where the various delimiter characters appeared
in the fields.

The correct algorithm is to enclose a field in the field enclosure
character if the field contains any special character. In addition
the field enclosure character and the escape character must be
escaped with an escape character.

Also, the various delimiter characters were class member
variables. However, their values were not actually used to generate the SQL
so if they were ever changed then the code would not have worked.