Use the composer autoloader to load files.
This change switches the XDMoD code to be loaded with the composer
loader rather than the hybrid composer and require and custom loader.
The main reason for this change is to switch the ETL directory to use
a psr-4 loader thus allowing xdmod modules to extend the ETL framework.
Previously it was not possible for a module to add a new DataEndpoint
due to the way classes were loaded.

The Appkernels module still uses the linker.json so this code path
is not modified at this time.

Also there was an unused file classes/DB/DBModel/TableColumn.php
rather than add this to the loader path I just removed it.

Also I switched on the optimizer for composer.