Improve input validation for download_report controller
The download_report controller endpoint already checks that the
filename is acceptable. However, in the case of an invalid filename
it potentially would try to open a file called [TMPDIR]/.[SUFFIX].

This change add an explicit exit if an invalid filename is supplied.

Also add a test for this scenario and to check some of the other input
validation code.