Add default value for the report format validator.
If a default is not supplied then the validator returns false
if validation fails. The isValidFormat() function expects a string
and this leads to a warning if false is supplied (although the
isValidFormat correctly returns).

This change updates the default to be an empty string which is
not a valid format, but is the coreect data type for the isValidFormat()
function.