Update the way that expected output is determined for UsageChartsTest
Due to the fonts on Centos8 being slightly different than those on Centos7 we
needed a way for the tests to utilize OS Specific expected results files.

We first need to know what OS we're running on so I added an `$osInfo` member
variable that `UsageChartsTest` will attempt to populate with the contents of
`/etc/os-release` which is an ini file that contains information about the OS
that is currently installed.

Next I updated the `chartSettingsProvider` function to optionally look for an
OS specific expected file first, falling back to the default `expected.json`
file in the case that one is not found. I then reverted the contents of
`expected.json` to it's centos7 values and added a new centos8 version.