Ensure dashboard is updated whenever reports change.
Previously the daashboard was only updated when a report was deleted.
This change moves the request_refresh call to the callback for the
report generator report store. This ensures that whenever the store
is updated (chart add, deletes and edits) then the dashboard will
be updated.

Tested manually by adding, renaming and deleting reports and switching
back to the dashboard to confirm that it reloads. I also confirmed that the
dashboard does NOT reload when you just view a report and don't edit it.