Ensure "Don't show me this" checkbox is honoured
The "Don't show me this again" checkbox only did something if
the user clicked "no" to view the tour. However the user interface
gives no indication that the checkbox only works when you click no.

This changes the code so that the checkbox is honoured independently
of whether you click yes or no. The annoying popup that you get when
you click no is still shown.

I also couldn't help myself and removed some cruft - you don't need
to specify any parameters to the viewedUser endpoint.