Changes to allow for existing installations with empty username users
These two changes will ensure that upgrades for existing installations that have
managed to create a user with a null username will complete sucessfully instead
of erroring out when the ETL framework attempts to update the User table. Once
these null username Users have are converted to empty string username Users then
a SQL script will be run that deactivates the User as it should be possible to
create a user with an empty / null username.