Update Query Filtering to be based on acls again
This PR's purpose is to replace the hard coded logic of
`Parameters::getParameters`, the function that determines how a query should be
filtered based on what a users most privileged acl is, with a database query
that utilizes the `user_acl_group_by_parameters` table.

To accomplish this we first need to capture for each acl which, if any,
dimension filtering is required. Thankfully our configuration files contain this
information already in the `dimensions` property. A new table was created
`acl_dimensions` that serve as a relation table between `acls` and `group_bys` (
dimension is another name for a group_by, and seeing as we already have an
`acl_group_bys` table, `acl_dimensions` it is. ). To ensure this table is
populated correctly the `acl-config` script was updated to include this as
another managed table. Next, to ensure that when a user is created or modified
the appropriate records are created in the `user_acl_group_by_parameeters` table
the `XDUser::saveUser` function was updated to include the maintenance of these
records.

New component tests ( and artifacts ) have been added to exercise the
`Parameters::getParameters` function.