Doing away with the TreeView in place filtering
Ultimately I don't think we're going to want to filter the TreeView in place
due to the asynchronous nature of the loading. We can hide nodes that haven't
been shown yet, but it's proving difficult to show them after they've been
hidden.

Not only that, but viewing the subsequently filtered data is problematic, again
due to the asynchronous nature of the tree. A user can open each pipeline to see
which nodes matched underneath, but they'd need to click each one individually.
I attempted to add an 'expandAll' after the filter process but even with a
number of nodes hidden it takes a loooong time to finish ( upwards of 5 seconds
or more ).

I think that a more useful display for a search would be a table w/ a minimum of
two columns, one that's the path to the node / attribute that matches the search
term and the other being the actual value that matched. We could then maybe
provide some context actions of navigating directly to the node in the
TreeView. This would be relatively easy as the tree already has an `expandPath`
function.

There could also be a `Open Pipeline | Action` feature that opens the pipeline /
action in the Graph View.