Updating to include OS specific expected values
For some reason the values returned for both `finfo` and in some cases the
response mimetype differ between PHP5.4 / PHP7.2 & Centos7 / Centos8 for the
same input. To account for this difference I've added an `$osSpecificExpected`
array that contains the expected svg and xml data for both centos7 and centos8.
This is then used to construct the expected output for the `testExport`
function.

I also updated `testExport` so that it always expects one and only one correct
value for both mimetype and `finfo`.