Add rest documentation.
This includes the rest schema file containing docs for the controller interfaces
and the rest interfaces. Some of these are missing descriptions, but this at least
is a place holder to get going.

Also add a static html page that pulls in a 3rd party js library that auto-generates
sensible looking html output.