Handle MySQL server has gone away errors.
This change handles the scenario when there is a long delay
between two different log message calls (for the database logger).
This can happen when aggregating / ingesting large amounts of data
and the log level is set low.

The change is to attempt a single retry if the specific 'server has
gone away' error is seen. This shouldn't change the behaviour for any
other types of sql error. If the retry fails then it does not try again.

This was tested by changing the `/etc/my.cnf.d/server.cnf` file in a
docker instance and adding the following:

```ini
wait_timeout = 30
```

I then put a `sleep(60)` as the last line of the run_aggregation() function
in /usr/lib64/xdmod/aggregate_supremm.php and ran it. (I also had the EtlJournalHelper
patch https://github.com/ubccr/xdmod-supremm/pull/285 installed).

Confirmed that the full processing happens with this patch installed and
errors out with a
```json
{"success":false,"count":0,"total":0,"totalCount":0,"results":[],"data":[],"message":"SQLSTATE[HY000]: General error: 2006 MySQL server has gone away","code":"HY000"}
```
otherwise