Centos8 Docker Image changes
Added the `allow-plugins` section due to the latest composer version prompting
whether or not to trust the two plugins listed here.

This is the base centos8 Dockerfile that produces a Centos8 image w/ all of
XDMoD 10.0's requirements pre-installed. Serves as the base image for
Dockerfile.centos8.

This Dockerfile is meant to use the image produced by
Dockerfile.centos8.xdmod-base as a base. It will then take care of installing
XDMoD 10.0. Changes have been made to certain files used by this Dockerfile to
account for changes needed when running on Centos8.

Centos8 requires a new version of the chromedriver, as such this script needed
to be updated to detect what version of centos it's running on and to copy the
correct version of chromedriver in place.

Just a small change here to enable permissive mode. This was causing

This file wasn't present before but used in `bin/services`.

This script was updated so that `php-fpm` was included in the start / stop
commands. I also found in my testing that `pidof` doesn't actually exist in the
centos7 image ( I was able to add it by installing `sysvinit-tools` ) so I
re-factored the `nodaemon` command to test for `pidof` before using it.

Refactored the removal of the php-pear-MDB2* packages so that we first test if
they're installed before attempting to remove them. This ensures that when they
are not installed ( i.e. on Centos8 ) the script doesn't error out causing build
problems.

Updated the version of chromedriver to 97.0 as this is what's required on
centos8. Tested w/ centos7 successfully.