Switch image comparison algorithm for regression tests.
This changes the regression tests for the usage charts to use
the `average_hash` image comparison algorithm from the imagehash
python module.

There are still two sets of expected results one for Centos 7 and
one for Centos 8. The image hash changes do not impact the SVG
regression tests - so these are different and then hashes themselves
are not completely the same between 7 & 8.