CentOS8 / Rocky Linux XDMoD Docker (#1607)
* Centos8 Docker Image changes

Added the `allow-plugins` section due to the latest composer version prompting
whether or not to trust the two plugins listed here.

This is the base centos8 Dockerfile that produces a Centos8 image w/ all of
XDMoD 10.0's requirements pre-installed. Serves as the base image for
Dockerfile.centos8.

This Dockerfile is meant to use the image produced by
Dockerfile.centos8.xdmod-base as a base. It will then take care of installing
XDMoD 10.0. Changes have been made to certain files used by this Dockerfile to
account for changes needed when running on Centos8.

Centos8 requires a new version of the chromedriver, as such this script needed
to be updated to detect what version of centos it's running on and to copy the
correct version of chromedriver in place.

Just a small change here to enable permissive mode. This was causing

This file wasn't present before but used in `bin/services`.

This script was updated so that `php-fpm` was included in the start / stop
commands. I also found in my testing that `pidof` doesn't actually exist in the
centos7 image ( I was able to add it by installing `sysvinit-tools` ) so I
re-factored the `nodaemon` command to test for `pidof` before using it.

Refactored the removal of the php-pear-MDB2* packages so that we first test if
they're installed before attempting to remove them. This ensures that when they
are not installed ( i.e. on Centos8 ) the script doesn't error out causing build
problems.

Updated the version of chromedriver to 97.0 as this is what's required on
centos8. Tested w/ centos7 successfully.

* Checkpointing centos8_docker work

* Checkpointing

- Added some additional comments to help clarify things.
- Added a `composer install` step.

- Added a couple of comments
- Added a RUN step that changes up the repos to reflect the sunsetting of
  Centos8

* Adding some more documentation and fixing Centos8 repo url

Since CentOS8 is not supported anymore they changed the repo urls. This change
makes it so that we can still build these images.

* Adding Docker Files for Rocky Linux

- Added first go at Rocky Linux docker images
- Updates to bin/services & copy-caches to account for new `VERSION_ID` in Rocky
  Linux ( 8.5 ).

* Updating source image for Rocky8 Docker

* simplifying pkg removal

* Updating centos8 image

* fixing centos8 image name

* Adding rocky8 to CircleCI builds

* fixing rocky8 addition to circleci

* Keeping Rocky8 executor in but removing it from the build matrix

* Adding regression images for rockylinux

These new regression images will enable our regression tests to run on
RockyLinux.

* Updates to support centos8 / rockylinux

* Making expected-rocky8.5.json posix compliant

* Adding Rocky executor back to the CircleCI pipeline

* Adding the rocky8-executor to the custom ui test setup

* Update tests/ci/Docker/Dockerfile.centos8.xdmod-base

Being more specific with which Centos8 image we depend on.

Co-authored-by: Jeffrey T. Palmer <jeffrey.t.palmer@gmail.com>

* Updates per code review by @jtpalmer

- Dockerfile.centos8:
  - removed the `mysql-prepare-db-dir` file as it's already present.
  - Updated an empty comment
  - Removed the unnecessary step of installing the UI test dependencies. This is
    handled later on by the `runtests.sh` script.
- Dockerfile.centos8.xdmod-base:
  - Updated the files that the sed scripts operate on as they were changed.
  - Removed the duplicate sed script.
  - Removed: `gmp-devel`, `java-1.8.0-openjdk`, and `java-1.8.0-openjdk-devel`
    as they are no longer required.
- Dockerfile.rocky8:
  - Updated the image used as the base for this Dockerfile as we've updated to
    using Rocky8.5 from Rocky8.
  - Removed the `mysql-prepare-db-dir` file as it's already present.
- Dockerfile.rocky8.xdmod-base:
  - Updated base image from Rocky8 to Rocky8.5
  - Removed: `gmp-devel`, `java-1.8.0-openjdk`, and `java-1.8.0-openjdk-devel`
    as they are no longer required.
- bin/mysql-prepare-db-dir:
  - This file is being removed as it is present in both CentOS8 & Rocky8/8.5
- bin/services:
  - The format for this file was updated to follow the original 4 space / tab.
- bootstrap.sh:
  - updated the files that the sed scripts operate on as they have changed.

* Installing UI test dependencies for CentOS/Rocky8

* Making bootstrap.sh posix compliant

* nonesense commit to appease the qa gods

* removing additional line

* Updating the rocky-8.5 regression expected values per version update

* Update tests/ci/Docker/bin/services

formatting update

Co-authored-by: Jeffrey T. Palmer <jeffrey.t.palmer@gmail.com>

* Update tests/ci/bootstrap.sh

Removing extraneous letter

Co-authored-by: Jeffrey T. Palmer <jeffrey.t.palmer@gmail.com>

* Removing CentOS8 repo changes as theyve been backed in

* Updating CentOS/Rocky images to latest

* Removing OS Specific Code

All of the docker images now use the same version of chromium now so we can
re-simplify this section of code.

* Appeasing the qa gods

* Updating build steps for CentOS8

* lksdfj

* Updating Rocky8 Image

* Removing a now unused variable

this variable is no longer used and can be safely deleted.

* Reverting copy-caches

Co-authored-by: Ryan Rathsam <build@shippable.com>
Co-authored-by: Jeffrey T. Palmer <jeffrey.t.palmer@gmail.com>