Centralizing the setting and usage of baseURL
These changes make it so that the baseURL only has to be set in one
place, `global-setup.ts` and the rest of the test suite will use that
value when executing.

Further changes will probably include making this value derrived from an
environment variable so that it can be set at run time when the docker
containers are spun up.