Updating the datasetid float comparison (PHP7.2) (#1612) (#1652)
* Updating the datasetid float comparison

Comparing floats as strings is not desirable due to the way they are represented
internally. Per https://www.php.net/manual/en/language.types.float.php they
should be compared to within a certain number of digits. This was found by
investigating an error where the "Show Raw Data" context menu item was not
showing the associated jobs. It turns out that the id's being compared were the
following:
- 0.22694672549385309
- 0.22694672549385

The latter number having been truncated after 14 digits, which was a problem
as they were being compared as strings, not floats. By comparing them as floats
we can mitigate this type of rounding / truncating error in the future.

* Testing a change from random float to random it

* Style changes for CCR.js

* Adding the randomInt function

* reverting changes to get_rawdata

* Adding comment to fix extra checks

* Reverting CCR.js

Trying to revert all the style changes to CCR.js as we're not putting
the randomInt function here.

Co-authored-by: Ryan Rathsam <build@shippable.com>

Co-authored-by: Ryan Rathsam <build@shippable.com>