Adding mongodb dependencies to XDMoD (#1656)
* Adding mongodb deps for xdmod

* Updating centos8 image

* working on circleci ui tests

* Adding os specific composer files

With the divergence between CentOS7 and 8 in regards to how they handle
the mongodb driver installation ( via rpm in CentOS7 vs via pecl in 8)
in addition to which driver is actually installed ( pecl's mongo library
vs. pecl's mongodb library do to the mongo library being abandoned ), it
has become necessary, at least temporarily, that we maintain a different
composer.json file for el7 vs. el8 builds. This will only impact el8
installations that want to utilize the `xdmod-supremm` module, and only
in that there are two additional system requirements, a
`yum install php-devel php-pear` and a `pecl install mongodb`.

These changes necessitated a modification of the CircleCI build so that
the `COMPOSER` environment variable is set to point to the appropriate
composer-el[7|8].json file.

Another note is that these changes / additional composer.json file will
only be necessary until 10.5 when we discontinue support for CentOS7.

* Updating the CircleCI composer cache

By adding the OS specific composer.json files we now also need to
update the CircleCI composer caches to account for this.

* Updating centos8 composer deps

* Ensuring that QA tests use composer.json

* second times the charm?

* third times the charm

* Updating README.md to account for the split composer files

* Appeasing style gods

* less is more

Co-authored-by: Ryan Rathsam <build@shippable.com>