Adding latest-releases to our public docs
This adds a new intermediate `latest-releases` page to our public facing
documentation so that we can provide another touch point with the users
looking to install / get started with XDMoD. The impetous for this
addition was us needing to manage the multiple el8 beta releases for
XDMoD 10.0. The old link would take them directly to the original 10.0
release which had an out of date el8 rpm. The new structure will direct
them to the new intermediate page and then more accurately direct the
user to the specific release, el7 or el8 as is appropriate.