Update table definition for ReportTemplateCharts
There is no unique index on the report template charts which means that the
normal etlv2 ingestion code cannot be used to sync the template definitions
with the database. You would have to write custom SQL to update an existing
row. Also if you run the existing ingestor more than once you'll get more
than one copy of each chart in every report!

Also change the database charset and collation to be utf8mb4 so that non-latin
characters can appear in the templates.