Updating the request timing start time name
So it turns out that in v3.0.9 of symfony/http-foundation there is a
request attribute named `start` which overlapes the variable that was
being used to track request elapsed time. This was causing a subset of
JobViewer tests to fail. I've updated the attribute name to
`timing.start` and these tests now pass.