Improve legend text generation algorithm
Previously timeseries plots with multiple dataset would always append
the units for the legend text. This results in lots of redundant text
if multiple dataset with the same units are plotted.

This pull request changes the code to only show the units if it actually
is useful to show (i.e. if the units are different between the datasets or
if  the 'single y-axis' config setting is set).