Fix regression test runner.
The UsageChartsTest was not loading the expected results
file because the variable containing the filename was not
initalised when it needed to be used.

Leading to this message:
```
PHP Warning:  file_get_contents(): Filename cannot be empty in /root/xdmod/tests/regression/lib/Controllers/UsageChartsTest.php on line 119
```
and the tests all getting skipped.

This change fixes the problem so that the tests are actually
run.