Add helper script that runs chromium.
The feature that we were previously using in chromium_headless v110 and
earlier seems to have been removed in v111. This changes the code
to use puppeteer-core to interact with chromium.

This change is intended to be a drop in replacement: The portal_settings
config that specifies the path to chromium is still used and we package the
puppeteer code so no need for post install steps.