Add additional expected output artifacts
for the standard API token tests (empty token, malformed token,
etc.), and decide which HTTP codes and file names to use for the
tests based on whether the endpoint is a controller or a REST
endpoint and whether the token is required or optional for
authentication. Also remove unused 'use' statement and make
the token endpoint path a variable to avoid repeating the string.