Refactor
to use existing Configuration functionality for parsing JSON files
rather than implementing new '$ref' and 'extends' keys. Also:
- Move the `TokenHelper` into a new class `TokenAuthTest` that
can be extended by a test instance rather than having to pass the
test instance as a parameter.
- Rename `makeRequest` method to `requestAndValidateJson`
and have it read the endpoint information from an test artifact file.
- Rather than using a `$validationType` variable, use the fact that
the validation object contains a `'$schema'` key to determine
that it is a JSON schema.
- Group together test output artifact files into single files used by
the tests that use them.
- Add comments to help with readability.