Refactor based on testing other PRs:
- Allow `runTokenAuthTests()` to run multiple successful tests as
defined in JSON input/output test artifacts.
- When running a `runTokenAuthTest()`, make one request with
the token in the header and one with it only in the query
parameters to ensure the result is the same.
- Instead of transforming fragment-only references, just ignore
them and let the JSON schema validator handle them.
- Add `"$id"` keys to the JSON schemas so that the validator
can properly process `"$ref"` keys if the object is in an `allOf`
statement.
- Rename `success.json` to `success_body.json` to be
consistent with other names.
- For test assertion messages, include both the path to where the
JSON artifact is defined and the (truncated) text of the artifact
itself.