Fix job array display.
The current code has incomplete support for job arrays - only the
primary job identifier is shown. This change adds a new internal parameter
called provider_job_id which contains the `job index` for non-array
jobs and the `base_job_index[array_index]` for array jobs.

There is a corresponding pull request for the xmdod-supremm module
and for the xdmod-xsede module that should all be merged at the same
time.