Refactor
to generate tokens that can be used by all tests, and
store them statically so that they do not have to be
generated more than once. Also require the input
and output test groups and filenames to be the
same for simplicity. Also fix bug in which the 'cd' role
was not included in the base roles.