Fixes for warnings generated with the supremm module tests
The xdmod-supremm module build has been updated to check for
errors/warnings. Needless to say I found some.

The default value for the font size must be a numeric string - and
not a text string.

Add some checks for data validity and hoy out an execption if the
checks fail.