Modify update script to use gsed if available.
The update script requires GNU sed (which is the default
sed on the RedHat Linuxes). OSX ships with a BSD sed which has
a different command syntax. This change results in the script preferring
the gsed command over sed if present. On an OSX machine you can install
GNU sed via homebrew: `brew install gsed`