Refactoring randomInt function location (#1724)
* Doc updates rocky8 (#1722)

* Updating the rocky8 latest release url

* Updated the Rocky 8 specific software reqs

Since the mongodb system level req has changed from being a yum
installed package to a pecl installed one those changes need to be
reflected in our documentation.

* Updates per code review by @connersaeli

* Refactoring randomInt function location

Moving this function so that it can be used by other areas of code not
just the AddDataPanel. This is to specifically address an issue with
looking up jobs via the Efficiency Tab.