Update XDUser.php
It seems like you're trying to set the value of the password_last_updated field in an associative array using the NOW() function. However, NOW() is a SQL function that returns the current date and time, and you're using it as if it's a PHP function.

If you're trying to set the current date and time in a PHP associative array, you should use PHP's built-in date() function like this:

$update_data['password_last_updated'] = date('Y-m-d H:i:s');
This will generate a string in the format YYYY-MM-DD HH:MM:SS, which is the format typically used for date and time values in MySQL databases.

Also attempting to set the organization_id field in an associative array based on the value of $this->_organizationID. However, using the string 'NULL' or the actual integer value depending on whether $this->_organizationID is null is correct for an SQL query.

However, since you're directly setting the value within a PHP associative array, you should use the actual null value (without quotes) when $this->_organizationID is null.

This way, you're setting the organization_id to null when $this->_organizationID is null, and to the actual integer value when $this->_organizationID has a value. This is the correct way to handle NULL values in PHP arrays that will later be used in SQL queries.