Refactor some of the integration tests
to include test data in PHP associative arrays rather than
in separate test artifact JSON files, to make it easier to
write, debug, and maintain the tests, and to make the
tests complete faster. Also only run the token
authentication tests for a single public user and a single
non-public user, to make the tests complete faster.