Add additional information for 11.0 upgrade
It's important for installations that host the XDMoD DB on a different
server than the XDMoD Web Server to ensure that the database user XDMoD
uses is updated to reflect a change in the XDMoD Web Server's hostname.
Failure to do this will result in XDMoD not being able to funciton due
to permission denied errors.