Adding ability to save UsageChart images (#1973)
These changes add the ability to, by specifying a env variable, to
instruct the UsageChart image export tests to save the actual images to
the `/tmp/` directory so that they can be compared to another set of
images to determine why these tests may be failing.

These changes will also output a metadata file to be used in further
determining potential similiarities in test failures.