Checkpointing Test Fixes
The big changes here are related to SSO & the internal dashboard. The
dashboard changes just revolved around making sure that the urls are
consistent on the backend & frontend.

The SSO changes implemented SAML SSO, tested both with the nodejs saml
test project and keycloak ( I'm not including keycloak in these changes
as that should be added in it's own PR ). At the moment I've used a
symfony bundle for SAML authentication nbgrp_one_login, but now that I
think about it really should be simplesamlphp. If only to minimize the
amount of change. That will be the next set of changes committed.