Getting SSO working with returnTo params
These changes get SSO working with the `saml-idp` nodejs app. Now that
we're testing xdmod from a box that doesn't have it installed, knowing
when a particular parameter is going to be used on a host that doesn't
have XDMOD installed is important. A bunch of these changes are to make
sure things are consistent.