Logging Updates to prepare for PHP 8.2
These updates are to prepare for the PHP 8.2 updates / Symfony Migration as the
new version of Monolog we're updating to no longer allows us to provide arrays
as messages. Thankfully the new version of Monolog does allow for classes that
implement `\Stringable` to be provided as a message. To take advantage of this I
created a new class called `LogOutput` that implements `\Stringable` and takes /
provides as a public member variable an array. This array is used when
`__toString()` is called and outputs information as we expect when logging to
the console. When logging to the database we instead expect to output
json_encoded data. To accomplish this I've added a new "Processor" ( right now
this is just an anonymous function, in the Symfony Migration they have a class
for these ) to the `CCRDBHandler` that will "enrich" the record being logged by
setting a new property called 'formatted' (for the new version of Monolog)
prior to CCRDBHandler writing the record to the db.

The changes were originally included in the Symfony Migration PR
but they don't actually require Symfony to function and will make the Symfony
PR smaller.