Playwright updates for xdmod-supremm (#2027)
* Playwright updates for xdmod-supremm

These minor changes were found while getting the playwright tests
working for xdmod-supremm.

* Updates per code review by @jpwhite4

* removing unused import

* Adding additional changes from xdmod-supremm

* Changing how we wait for mask to be gone

These changes update the way we wait for the mask to not be present from
waiting for it to not be visible to waiting for the locator to return 0
( i.e. that it's not in the dom ).

I also included an update from 1.51.1 to 1.54.1 for playwright

* Missed a this.mask change

* Updating to Rocky 8.10

---------

Co-authored-by: Joe White <jpwhite4@buffalo.edu>