Increase reliability of Playwright mask detection
The previous mask detection code was based on the assumption that when masks
were "un-masked" they would always be removed from the dom, this is not the case
with the Job Viewer as it is instead hidden from view. This caused the original
selector to return more than one element, which caused the `toHaveCount(0)` to
fail, even though there were no masks visible to the user. The new method relies
on detecting a side effect of the masking process, that being that when any mask
is created the `x-masked` class is added to the `body` dom element and is
subsequenetly removed when the mask "un-masked". This happens regardless of
whether or not the mask is removed from the dom or hidden. It comes with the
side benefit of testing an element that there can only be one of on any given
page, so we don't run into the same "multiple elements found" error case.

The old metric explorer dialog selector did not test for visibility, which
caused problems when other elements matched the selector resulting in more than
one element being returned.