Fix formatting issues with the logger.
This fixes two bugs in the output format for the logger:
1) The line based formatter had incorrect spacing between the message
   and context information (when present)
2) The line based formatter was ommiting the stack trace when
   printing exceptions
3) The database handler was not configured to output the context information

I've added some component tests to verify the correct output. The reference
values for the output were generated on the old code.

Known issues:

The serialization of exceptions is different in this version of monolog vs
the XDMoD 10.5 version. The old code would print exceptions like:

```
2025-10-29 15:46:44 [error]  (exceptiontest: Exception: Test Line Exception in /root/xdmod/tests/component/lib/LoggerTest.php:16
Stack trace:
```

and the new code does it like this:

```
2025-10-29 15:17:02 [error] (exceptiontest: [object] (Exception(code: 0): Test Line Exception at /root/xdmod/tests/component/lib/LoggerTest.php:16)
[stacktrace]
)
```

I'm not going to make the code use the old exception serialization format at this time
I think it only impacts people reading the logs and the new format provides more information
than the old one.