Updates to httpd.conf for Playwright (#2120)
Because Playwright runs in a different docker container than xdmod it
needs to be able to make requests to the xdmod container host.
Previously we made changes to the xdmod containers httpd.conf file and
restarted the service before running the Playwright tests. By updating
the httpd.conf template we no longer need these extra CircleCI steps to
allow the Playwright tests to run.

I have also added a test to `tests/ci/validate.sh` that checks that the
xdmod containers webserver responds to `xdmod`.