Fix errors in XDEBUG install
This updates the xdebug log output file location to /var/log/xdmod
so that it will be stored in the CI artifacts (and hence available
to us after a CI build).

Also create the profile and coverage directories and set appropriate
permissions so they are writable by processes run by apache or xdmod.

Also disable generation of profile information by default. This is
not used by the CI builds so no need to have it on.